PI-1050 Return and sort by raw Elasticsearch score (#561)

+++ src/main/kotlin/uk/gov/justice/hmpps/probationsearch/addresssearch/AddressSearch.kt
@@ -59,7 +59,7 @@ data class AddressSearchResponses(
 data class AddressSearchResponse(
   val person: Person,
   val address: Address,
-  val matchScore: Int,
+  val matchScore: Float,
 )
 
 data class PersonDetail(
